refactor(gpu): LOG2-first domain-size constants and u32 log2 typing across the GPU stack (#349)
## What ❔

Stacked on #348 (retarget to `av_gkr_compiler` after it merges).

- Invert `circuit_type.rs` domain-size constants to LOG2-first: read the
upstream `DOMAIN_SIZE_LOG2` trait consts directly and compute sizes as
`1 << log2`, replacing the backwards `trailing_zeros()` derivations
(incl. one in production `config.rs`).
- Type all log2 values as `u32` end-to-end inside the GPU stack
(`get_domain_size_log2()` on all five circuit-type enums, ~19
production/test params like `final_trace_size_log_2` / `trace_len_log2`
/ `query_domain_log2`, the `execution_prover` origin, test consts).
Casts remain only at genuine boundaries: upstream CPU APIs and
`ProverConfig` usize fields, `gpu_gkr_model`'s host-side layout API, and
slice indexing.

## Why ❔

The log2 authorities are `u32` (trait consts, `trailing_zeros()`), and
GPU kernel params consume `u32` — the previous `usize` threading forced
u32 → usize → u32 roundtrips (e.g. 9 casts in `whir/mod.rs` alone).
Values are provably unchanged; the type now flows forward from the
authority.

Validated: lib + test-target +
`gpu_execution_prover`/`gpu_program_prover` compile; unit tests green;
`run_jump_branch_slt_proof_parity_test` end-to-end on GPU (bit-identical
proof vs CPU) at this exact tree.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted.

+++ gpu/circuit_prover/src/prover/gkr/backward/main_layer/extras.rs
@@ -227,8 +227,8 @@ where
 pub(crate) fn derive_dimension_reducing_inputs(
     initial_layer_idx: usize,
     initial_output_map: &BTreeMap<OutputType, Vec<GKRAddress>>,
-    initial_trace_log_2: usize,
-    final_trace_log_2: usize,
+    initial_trace_log_2: u32,
+    final_trace_log_2: u32,
 ) -> BTreeMap<usize, BTreeMap<OutputType, DimensionReducingInputOutput>> {
     let mut result: BTreeMap<usize, BTreeMap<OutputType, DimensionReducingInputOutput>> =
         BTreeMap::new();
